+++ code/game/machinery/bots/cleanbot.dm
@@ -4,11 +4,11 @@
 	name = "proxy bucket"
 	icon = 'icons/obj/aibots.dmi'
 	icon_state = "bucket_proxy"
-	force = 3.0
-	throwforce = 5.0
+	force = 3
+	throwforce = 5
 	throw_speed = 2
 	throw_range = 5
-	w_class = 3.0
+	w_class = 3.
 	var/created_name = "Cleanbot"
 
 
@@ -18,7 +18,7 @@
 	desc = "A little cleaning robot, he looks so excited!"
 	icon = 'icons/obj/aibots.dmi'
 	icon_state = "cleanbot0"
-	layer = 5.0
+	layer = 5
 	density = 0
 	anchored = 0
 	//weight = 1.0E7
@@ -27,7 +27,6 @@
 	var/blood = 1
 	var/list/target_types = list()
 	var/obj/effect/decal/cleanable/target
-	var/obj/effect/decal/cleanable/oldtarget
 	var/max_targets = 50 //Maximum number of targets a cleanbot can ignore.
 	var/oldloc = null
 	req_one_access = list(access_janitor, access_robotics)
@@ -40,7 +39,7 @@
 	var/next_dest_loc
 	radio_frequency = SERV_FREQ //Service
 	bot_type = CLEAN_BOT
-	bot_filter = RADIO_CLEANBOT
+	model = "Cleanbot"
 
 /obj/machinery/bot/cleanbot/New()
 	..()
@@ -48,11 +47,8 @@
 	icon_state = "cleanbot[on]"
 
 	var/datum/job/janitor/J = new/datum/job/janitor
-	botcard.access = J.get_access()
+	botcard.access += J.get_access()
 	prev_access = botcard.access
-
-	spawn(5)
-		add_to_beacons(bot_filter)
 
 /obj/machinery/bot/cleanbot/turn_on()
 	..()
@@ -68,7 +64,6 @@
 	..()
 	ignore_list = list() //Allows the bot to clean targets it previously ignored due to being unreachable.
 	target = null
-	oldtarget = null
 	oldloc = null
 
 /obj/machinery/bot/cleanbot/set_custom_texts()
@@ -76,14 +71,14 @@
 	text_dehack = "[name]'s software has been reset!"
 	text_dehack_fail = "[name] does not seem to respond to your repair code!"
 
-/obj/machinery/bot/cleanbot/attack_hand(mob/user as mob)
+/obj/machinery/bot/cleanbot/attack_hand(mob/user)
 	. = ..()
 	if (.)
 		return
 	usr.set_machine(src)
 	interact(user)
 
-/obj/machinery/bot/cleanbot/interact(mob/user as mob)
+/obj/machinery/bot/cleanbot/interact(mob/user)
 	var/dat
 	dat += hack(user)
 	dat += text({"
@@ -109,13 +104,9 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 			blood =!blood
 			get_targets()
 			updateUsrDialog()
-		if("freq")
-			var/freq = text2num(input("Select frequency for  navigation beacons", "Frequency", num2text(beacon_freq / 10))) * 10
-			if (freq > 0)
-				beacon_freq = freq
 			updateUsrDialog()
 
-/obj/machinery/bot/cleanbot/attackby(obj/item/weapon/W, mob/user as mob)
+/obj/machinery/bot/cleanbot/attackby(obj/item/weapon/W, mob/user, params)
 	if (istype(W, /obj/item/weapon/card/id)||istype(W, /obj/item/device/pda))
 		if(allowed(user) && !open && !emagged)
 			locked = !locked
@@ -130,13 +121,13 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 	else
 		return ..()
 
-/obj/machinery/bot/cleanbot/Emag(mob/user as mob)
+/obj/machinery/bot/cleanbot/Emag(mob/user)
 	..()
 	if(emagged == 2)
 		if(user)
 			user << "<span class='danger'>[src] buzzes and beeps.</span>"
 
-/obj/machinery/bot/cleanbot/process_scan(var/obj/effect/decal/cleanable/D)
+/obj/machinery/bot/cleanbot/process_scan(obj/effect/decal/cleanable/D)
 	for(var/T in target_types)
 		if(istype(D, T))
 			return D
@@ -157,32 +148,25 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 
 			if(prob(5)) //Spawns foam!
 				visible_message("<span class='danger'>[src] whirs and bubbles violently, before releasing a plume of froth!</span>")
-				new /obj/effect/effect/foam(loc)
+				PoolOrNew(/obj/effect/particle_effect/foam, loc)
 
 	else if (prob(5))
-		visible_message("[src] makes an excited beeping booping sound!")
+		audible_message("[src] makes an excited beeping booping sound!")
 
 	if(!target) //Search for cleanables it can see.
-		target = scan(/obj/effect/decal/cleanable/, oldtarget)
-		oldtarget = target
+		target = scan(/obj/effect/decal/cleanable/)
 
-	if(!target)
-		if(loc != oldloc)
-			oldtarget = null
+	if(!target && auto_patrol) //Search for cleanables it can see.
+		if(mode == BOT_IDLE || mode == BOT_START_PATROL)
+			start_patrol()
 
-		if(auto_patrol)
-			if(mode == BOT_IDLE || mode == BOT_START_PATROL)
-				start_patrol()
-
-			if(mode == BOT_PATROL)
-				bot_patrol()
-
-		return
+		if(mode == BOT_PATROL)
+			bot_patrol()
 
 	if(target)
 		if(!path || path.len == 0) //No path, need a new one
 			//Try to produce a path to the target, and ignore airlocks to which it has access.
-			path = get_path_to(loc, target.loc, /turf/proc/AdjacentTurfsWithAccess, /turf/proc/Distance, 0, 30, id=botcard)
+			path = get_path_to(loc, target.loc, src, /turf/proc/Distance, 0, 30, id=botcard)
 			if (!bot_move(target))
 				add_to_ignore(target)
 				target = null
@@ -204,14 +188,28 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 /obj/machinery/bot/cleanbot/proc/get_targets()
 	target_types = new/list()
 
-	target_types += /obj/effect/decal/cleanable
+	target_types += /obj/effect/decal/cleanable/oil
+	target_types += /obj/effect/decal/cleanable/vomit
+	target_types += /obj/effect/decal/cleanable/robot_debris
+	target_types += /obj/effect/decal/cleanable/crayon
+	target_types += /obj/effect/decal/cleanable/molten_item
+	target_types += /obj/effect/decal/cleanable/tomato_smudge
+	target_types += /obj/effect/decal/cleanable/egg_smudge
+	target_types += /obj/effect/decal/cleanable/pie_smudge
+	target_types += /obj/effect/decal/cleanable/flour
+	target_types += /obj/effect/decal/cleanable/ash
+	target_types += /obj/effect/decal/cleanable/greenglow
+	target_types += /obj/effect/decal/cleanable/dirt
 
 	if(blood)
-		target_types += /obj/effect/decal/cleanable/blood
-		// target_types += /obj/effect/decal/cleanable/blood/gibs/ //Already subtypes
-		// target_types += /obj/effect/decal/cleanable/blood/trail_holder
+		target_types += /obj/effect/decal/cleanable/xenoblood/
+		target_types += /obj/effect/decal/cleanable/xenoblood/xgibs
+		target_types += /obj/effect/decal/cleanable/blood/
+		target_types += /obj/effect/decal/cleanable/blood/gibs/
+		target_types += /obj/effect/decal/cleanable/blood/drip/
+		target_types += /obj/effect/decal/cleanable/trail_holder
 
-/obj/machinery/bot/cleanbot/proc/clean(var/obj/effect/decal/cleanable/target)
+/obj/machinery/bot/cleanbot/proc/clean(obj/effect/decal/cleanable/target)
 	anchored = 1
 	icon_state = "cleanbot-c"
 	visible_message("<span class='notice'>[src] begins to clean up [target]</span>")
@@ -226,7 +224,7 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 
 /obj/machinery/bot/cleanbot/explode()
 	on = 0
-	visible_message("<span class='userdanger'>[src] blows apart!</span>")
+	visible_message("<span class='boldannounce'>[src] blows apart!</span>")
 	var/turf/Tsec = get_turf(src)
 
 	new /obj/item/weapon/reagent_containers/glass/bucket(Tsec)
@@ -236,16 +234,16 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 	if (prob(50))
 		new /obj/item/robot_parts/l_arm(Tsec)
 
-	var/datum/effect/effect/system/spark_spread/s = new /datum/effect/effect/system/spark_spread
+	var/datum/effect_system/spark_spread/s = new /datum/effect_system/spark_spread
 	s.set_up(3, 1, src)
 	s.start()
-	qdel(src)
-	return
+	..() //qdels us and removes us from processing objects
 
-/obj/item/weapon/bucket_sensor/attackby(var/obj/item/W, mob/user as mob)
+/obj/item/weapon/bucket_sensor/attackby(obj/item/W, mob/user as mob, params)
 	..()
 	if(istype(W, /obj/item/robot_parts/l_arm) || istype(W, /obj/item/robot_parts/r_arm))
-		user.drop_item()
+		if(!user.unEquip(W))
+			return
 		qdel(W)
 		var/turf/T = get_turf(loc)
 		var/obj/machinery/bot/cleanbot/A = new /obj/machinery/bot/cleanbot(T)
