+++ code/game/machinery/bots/ed209bot.dm
@@ -3,7 +3,7 @@
 	desc = "A security robot.  He looks less than thrilled."
 	icon = 'icons/obj/aibots.dmi'
 	icon_state = "ed2090"
-	layer = 5.0
+	layer = 5
 	density = 1
 	anchored = 0
 //	weight = 1.0E7
@@ -35,15 +35,7 @@
 	var/shoot_sound = 'sound/weapons/Taser.ogg'
 	radio_frequency = SEC_FREQ
 	bot_type = SEC_BOT
-	bot_filter = RADIO_SECBOT
-
-	//List of weapons that secbots will not arrest for
-	var/safe_weapons = list(\
-		/obj/item/weapon/gun/energy/laser/bluetag,\
-		/obj/item/weapon/gun/energy/laser/redtag,\
-		/obj/item/weapon/gun/energy/laser/practice,\
-		/obj/item/weapon/melee/truncheon/telebaton,\
-		/obj/item/weapon/gun/energy/kinetic_accelerator)
+	model = "ED-209"
 
 
 /obj/item/weapon/ed209_assembly
@@ -67,11 +59,9 @@
 	set_weapon() //giving it the right projectile and firing sound.
 	spawn(3)
 		var/datum/job/detective/J = new/datum/job/detective
-		botcard.access = J.get_access()
+		botcard.access += J.get_access()
 		prev_access = botcard.access
 
-
-		add_to_beacons(bot_filter)
 		if(lasercolor)
 			shot_delay = 6//Longer shot delay because JESUS CHRIST
 			check_records = 0//Don't actively target people set to arrest
@@ -108,7 +98,7 @@
 	text_dehack = "You restore [name]'s combat inhibitor."
 	text_dehack_fail = "[name] ignores your attempts to restrict him!"
 
-/obj/machinery/bot/ed209/attack_hand(mob/user as mob)
+/obj/machinery/bot/ed209/attack_hand(mob/user)
 	. = ..()
 	if (.)
 		return
@@ -128,15 +118,12 @@ Maintenance panel panel is [open ? "opened" : "closed"]<BR>"},
 Arrest Unidentifiable Persons: []<BR>
 Arrest for Unauthorized Weapons: []<BR>
 Arrest for Warrant: []<BR>
-<BR>
-Operating Mode: []<BR>
 Report Arrests[]<BR>
 Auto Patrol[]"},
 
 "<A href='?src=\ref[src];operation=idcheck'>[idcheck ? "Yes" : "No"]</A>",
 "<A href='?src=\ref[src];operation=weaponscheck'>[weaponscheck ? "Yes" : "No"]</A>",
 "<A href='?src=\ref[src];operation=ignorerec'>[check_records ? "Yes" : "No"]</A>",
-"<A href='?src=\ref[src];operation=switchmode'>[arrest_type ? "Detain" : "Arrest"]</A>",
 "<A href='?src=\ref[src];operation=declarearrests'>[declare_arrests ? "Yes" : "No"]</A>",
 "<A href='?src=\ref[src];operation=patrol'>[auto_patrol ? "On" : "Off"]</A>" )
 	var/datum/browser/popup = new(user, "autoed209", "Automatic Security Unit v2.6")
@@ -162,14 +149,11 @@ Auto Patrol[]"},
 		if ("ignorerec")
 			check_records = !check_records
 			updateUsrDialog()
-		if ("switchmode")
-			arrest_type = !arrest_type
-			updateUsrDialog()
 		if("declarearrests")
 			declare_arrests = !declare_arrests
 			updateUsrDialog()
 
-/obj/machinery/bot/ed209/attackby(obj/item/weapon/W as obj, mob/user as mob)
+/obj/machinery/bot/ed209/attackby(obj/item/weapon/W, mob/user, params)
 	if (istype(W, /obj/item/weapon/card/id)||istype(W, /obj/item/device/pda))
 		if (allowed(user) && !open && !emagged)
 			locked = !locked
@@ -195,7 +179,7 @@ Auto Patrol[]"},
 						shootAt(user)
 					mode = BOT_HUNT
 
-/obj/machinery/bot/ed209/Emag(mob/user as mob)
+/obj/machinery/bot/ed209/Emag(mob/user)
 	..()
 	if(emagged == 2)
 		if(user)
@@ -205,6 +189,19 @@ Auto Patrol[]"},
 		declare_arrests = 0
 		icon_state = "[lasercolor]ed209[on]"
 		set_weapon()
+
+/obj/machinery/bot/ed209/bullet_act(obj/item/projectile/Proj)
+	if(istype(Proj ,/obj/item/projectile/beam)||istype(Proj,/obj/item/projectile/bullet))
+		if((Proj.damage_type == BURN) || (Proj.damage_type == BRUTE))
+			if (!Proj.nodamage && Proj.damage < src.health)
+				if(ismob(Proj.firer))
+					var/mob/M = Proj.firer
+					threatlevel = M.assess_threat(src)
+					threatlevel += 6
+					if(threatlevel >= 4)
+						target = M
+						mode = BOT_HUNT
+	..()
 
 /obj/machinery/bot/ed209/bot_process()
 	if (!..())
@@ -255,8 +252,7 @@ Auto Patrol[]"},
 						icon_state = "[lasercolor]ed209[on]"
 					var/mob/living/carbon/M = target
 					if(istype(M, /mob/living/carbon/human))
-						if( M.stuttering < 5 && !(HULK in M.mutations) )
-							M.stuttering = 5
+						M.stuttering = 5
 						M.Stun(5)
 						M.Weaken(5)
 					else
@@ -268,7 +264,7 @@ Auto Patrol[]"},
 						var/area/location = get_area(src)
 						speak("[arrest_type ? "Detaining" : "Arresting"] level [threatlevel] scumbag <b>[target]</b> in [location].", radio_frequency)
 					target.visible_message("<span class='danger'>[src] has stunned [target]!</span>",\
-											"<span class='userdanger'>[src] has stunned [target]!</span></span>")
+											"<span class='userdanger'>[src] has stunned you!</span>")
 
 					mode = BOT_PREP_ARREST
 					anchored = 1
@@ -298,13 +294,13 @@ Auto Patrol[]"},
 						mode = BOT_ARREST
 						playsound(loc, 'sound/weapons/cablecuff.ogg', 30, 1, -2)
 						target.visible_message("<span class='danger'>[src] is trying to put zipties on [target]!</span>",\
-											"<span class='userdanger'>[src] is trying to put zipties on [target]!</span>")
+											"<span class='userdanger'>[src] is trying to put zipties on you!</span>")
 
 						spawn(60)
 							if( !Adjacent(target) || !isturf(target.loc) ) //if he's in a closet or not adjacent, we cancel cuffing.
 								return
 							if(!target.handcuffed)
-								target.handcuffed = new /obj/item/weapon/handcuffs/cable/zipties/used(target)
+								target.handcuffed = new /obj/item/weapon/restraints/handcuffs/cable/zipties/used(target)
 								target.update_inv_handcuffed(0)	//update the handcuffs overlay
 								back_to_idle()
 					else
@@ -393,22 +389,13 @@ Auto Patrol[]"},
 			continue
 
 /obj/machinery/bot/ed209/proc/check_for_weapons(var/obj/item/slot_item)
-	if(istype(slot_item, /obj/item/weapon/gun) || istype(slot_item, /obj/item/weapon/melee))
-		if(!(slot_item.type in safe_weapons))
-			return 1
+	if(slot_item && slot_item.needs_permit)
+		return 1
 	return 0
-
-/* terrible
-/obj/machinery/bot/ed209/Bumped(atom/movable/M as mob|obj)
-	spawn(0)
-		if (M)
-			var/turf/T = get_turf(src)
-			M:loc = T
-*/
 
 /obj/machinery/bot/ed209/explode()
 	walk_to(src,0)
-	visible_message("<span class='userdanger'>[src] blows apart!</span>")
+	visible_message("<span class='boldannounce'>[src] blows apart!</span>")
 	var/turf/Tsec = get_turf(src)
 
 	var/obj/item/weapon/ed209_assembly/Sa = new /obj/item/weapon/ed209_assembly(Tsec)
@@ -445,12 +432,12 @@ Auto Patrol[]"},
 			if(lasercolor == "r")
 				new /obj/item/clothing/suit/redtag(Tsec)
 
-	var/datum/effect/effect/system/spark_spread/s = new /datum/effect/effect/system/spark_spread
+	var/datum/effect_system/spark_spread/s = new /datum/effect_system/spark_spread
 	s.set_up(3, 1, src)
 	s.start()
 
-	new /obj/effect/decal/cleanable/blood/oil(loc)
-	qdel(src)
+	new /obj/effect/decal/cleanable/oil(loc)
+	..() //qdels us and removes us from processing objects
 
 /obj/machinery/bot/ed209/proc/set_weapon()  //used to update the projectile type and firing sound
 	shoot_sound = 'sound/weapons/laser.ogg'
@@ -468,7 +455,7 @@ Auto Patrol[]"},
 		else if(lasercolor == "r")
 			projectile = /obj/item/projectile/lasertag/redtag
 
-/obj/machinery/bot/ed209/proc/shootAt(var/mob/target)
+/obj/machinery/bot/ed209/proc/shootAt(mob/target)
 	if(lastfired && world.time - lastfired < shot_delay)
 		return
 	lastfired = world.time
@@ -495,7 +482,7 @@ Auto Patrol[]"},
 	A.xo = U.x - T.x
 	A.fire()
 
-/obj/machinery/bot/ed209/attack_alien(var/mob/living/carbon/alien/user as mob)
+/obj/machinery/bot/ed209/attack_alien(mob/living/carbon/alien/user)
 	..()
 	if (!isalien(target))
 		target = user
@@ -514,7 +501,7 @@ Auto Patrol[]"},
 		pulse2.anchored = 1
 		pulse2.dir = pick(cardinal)
 		spawn(10)
-			pulse2.delete()
+			qdel(pulse2)
 		var/list/mob/living/carbon/targets = new
 		for (var/mob/living/carbon/C in view(12,src))
 			if (C.stat==2)
@@ -542,7 +529,7 @@ Auto Patrol[]"},
 
 
 
-/obj/item/weapon/ed209_assembly/attackby(obj/item/weapon/W as obj, mob/user as mob)
+/obj/item/weapon/ed209_assembly/attackby(obj/item/weapon/W, mob/user, params)
 	..()
 
 	if(istype(W, /obj/item/weapon/pen))
@@ -555,7 +542,8 @@ Auto Patrol[]"},
 	switch(build_step)
 		if(0,1)
 			if(istype(W, /obj/item/robot_parts/l_leg) || istype(W, /obj/item/robot_parts/r_leg))
-				user.drop_item()
+				if(!user.unEquip(W))
+					return
 				qdel(W)
 				build_step++
 				user << "<span class='notice'>You add the robot leg to [src].</span>"
@@ -568,12 +556,15 @@ Auto Patrol[]"},
 					icon_state = "ed209_legs"
 
 		if(2)
+			var/newcolor = ""
 			if(istype(W, /obj/item/clothing/suit/redtag))
-				lasercolor = "r"
+				newcolor = "r"
 			else if(istype(W, /obj/item/clothing/suit/bluetag))
-				lasercolor = "b"
-			if(lasercolor || istype(W, /obj/item/clothing/suit/armor/vest))
-				user.drop_item()
+				newcolor = "b"
+			if(newcolor || istype(W, /obj/item/clothing/suit/armor/vest))
+				if(!user.unEquip(W))
+					return
+				lasercolor = newcolor
 				qdel(W)
 				build_step++
 				user << "<span class='notice'>You add the armor to [src].</span>"
@@ -587,7 +578,7 @@ Auto Patrol[]"},
 				if(WT.remove_fuel(0,user))
 					build_step++
 					name = "shielded frame assembly"
-					user << "<span class='notice'>You welded the vest to [src].</span>"
+					user << "<span class='notice'>You weld the vest to [src].</span>"
 		if(4)
 			switch(lasercolor)
 				if("b")
@@ -602,7 +593,8 @@ Auto Patrol[]"},
 					if(!istype(W, /obj/item/clothing/head/helmet))
 						return
 
-			user.drop_item()
+			if(!user.unEquip(W))
+				return
 			qdel(W)
 			build_step++
 			user << "<span class='notice'>You add the helmet to [src].</span>"
@@ -612,7 +604,8 @@ Auto Patrol[]"},
 
 		if(5)
 			if(isprox(W))
-				user.drop_item()
+				if(!user.unEquip(W))
+					return
 				qdel(W)
 				build_step++
 				user << "<span class='notice'>You add the prox sensor to [src].</span>"
@@ -624,10 +617,10 @@ Auto Patrol[]"},
 			if(istype(W, /obj/item/stack/cable_coil))
 				var/obj/item/stack/cable_coil/coil = W
 				if (coil.get_amount() < 1)
-					user << "<span class='warning'>You need one length of cable to wire the ED-209.</span>"
+					user << "<span class='warning'>You need one length of cable to wire the ED-209!</span>"
 					return
 				user << "<span class='notice'>You start to wire [src]...</span>"
-				if (do_after(user, 40))
+				if (do_after(user, 40, target = src))
 					if (coil.get_amount() >= 1 && build_step == 6)
 						coil.use(1)
 						build_step = 7
@@ -635,33 +628,36 @@ Auto Patrol[]"},
 						name = "wired ED-209 assembly"
 
 		if(7)
+			var/newname = ""
 			switch(lasercolor)
 				if("b")
 					if(!istype(W, /obj/item/weapon/gun/energy/laser/bluetag))
 						return
-					name = "bluetag ED-209 assembly"
+					newname = "bluetag ED-209 assembly"
 				if("r")
 					if(!istype(W, /obj/item/weapon/gun/energy/laser/redtag))
 						return
-					name = "redtag ED-209 assembly"
+					newname = "redtag ED-209 assembly"
 				if("")
 					if(!istype(W, /obj/item/weapon/gun/energy/gun/advtaser))
 						return
-					name = "taser ED-209 assembly"
+					newname = "taser ED-209 assembly"
 				else
 					return
+			if(!user.unEquip(W))
+				return
+			name = newname
 			build_step++
 			user << "<span class='notice'>You add [W] to [src].</span>"
 			item_state = "[lasercolor]ed209_taser"
 			icon_state = "[lasercolor]ed209_taser"
-			user.drop_item()
 			qdel(W)
 
 		if(8)
 			if(istype(W, /obj/item/weapon/screwdriver))
 				playsound(loc, 'sound/items/Screwdriver.ogg', 100, 1)
 				var/turf/T = get_turf(user)
-				user << "<span class='notice'>Now attaching the gun to the frame...</span>"
+				user << "<span class='notice'>You start attaching the gun to the frame...</span>"
 				sleep(40)
 				if(get_turf(user) == T)
 					build_step++
@@ -670,17 +666,18 @@ Auto Patrol[]"},
 
 		if(9)
 			if(istype(W, /obj/item/weapon/stock_parts/cell))
+				if(!user.unEquip(W))
+					return
 				build_step++
 				user << "<span class='notice'>You complete the ED-209.</span>"
 				var/turf/T = get_turf(src)
 				new /obj/machinery/bot/ed209(T,created_name,lasercolor)
-				user.drop_item()
 				qdel(W)
 				user.unEquip(src, 1)
 				qdel(src)
 
 
-/obj/machinery/bot/ed209/bullet_act(var/obj/item/projectile/Proj)
+/obj/machinery/bot/ed209/bullet_act(obj/item/projectile/Proj)
 	if(!disabled)
 		var/lasertag_check = 0
 		if((lasercolor == "b"))
